pyproject.toml
[project]
name = "m2-rag"
version = "0.1.0"
description = "A RAG-based chatbot for interacting with the M2 language."
requires-python = ">=3.12"
dependencies = [
  "faiss-cpu>=1.7.4",
  "minsearch>=0.0.7",
  "pydantic>=2.12.3",
  "pydantic-ai>=1.9.1",
  "openai>=2.8.1",
  "requests>=2.32.5",
  "sentence-transformers>=5.1.2",
  "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
  "pytest>=8.3.0",
]
